Rating missing in shop Filters not working for km and rating in Dashborad Screen

+++ lib/models/dashboard_models/near_by_shop_model.dart
@@ -30,7 +30,7 @@ class NearByShopModel {
   dynamic otpTime;
   String? vendorId;
   String? shopId;
-  double? rating;
+  int? rating;
   String? countryName;
   String? stateName;
   String? cityName;
